Fix contact page metadata by using client form component
