+++ src/components/ContactForm.tsx
@@ -1,0 +1,48 @@
+'use client';
+
+import { useState } from 'react';
+import { Button } from '@/components/ui/button';
+import { Input } from '@/components/ui/input';
+
+export default function ContactForm() {
+  const [form, setForm] = useState({ name: '', email: '', message: '' });
+
+  const handleChange = (
+    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
+  ) => {
+    setForm({ ...form, [e.target.name]: e.target.value });
+  };
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    // TODO: handle submit
+  };
+
+  return (
+    <form onSubmit={handleSubmit} className="max-w-xl mx-auto space-y-4">
+      <Input
+        name="name"
+        placeholder="Nome"
+        value={form.name}
+        onChange={handleChange}
+      />
+      <Input
+        type="email"
+        name="email"
+        placeholder="Email"
+        value={form.email}
+        onChange={handleChange}
+      />
+      <textarea
+        name="message"
+        className="w-full h-32 p-3 border rounded-md"
+        placeholder="Mensagem"
+        value={form.message}
+        onChange={handleChange}
+      />
+      <Button type="submit" className="w-full">
+        Enviar
+      </Button>
+    </form>
+  );
+}
